Map Slider Edits RD
This is a very rough draft of the system to filter the teams based on
year. However, I'm not sure if I referenced everything correctly, and
there is a missing piece of data (retirement year) which I have
temporarily substituted an iteger for. PLEASE feel free to debug and
correct what I missed. I apologize if this somehow makes the map
unusable in any way...
-JM